OneDrive/Desktop/smart-mess/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
qrcode[pil]==7.4.2
Pillow==10.3.0
python-multipart==0.0.9
